+++ src/main/java/com/example/demo/controller/PanCardController.java
@@ -20,6 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.demo.beans.PanCard;
 import com.example.demo.service.MapStateToError;
 import com.example.demo.service.PanCardService;
+import com.google.gson.Gson;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -27,7 +28,7 @@ import io.swagger.annotations.ApiParam;
 
 @RestController
 @RequestMapping(path = "/api/object/pan")
-@Api(value = "/api/object/pan", description = "PanCard Operation")
+@Api(value = "PanCard Crud Operations", description = "PanCard Crud Operation")
 public class PanCardController {
 
 	private static final Logger logger = LoggerFactory.getLogger(PanCardController.class);
@@ -40,42 +41,61 @@ public class PanCardController {
 
 	@RequestMapping(path = "/create", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
 	@ResponseBody
-	@ApiOperation(value = "/create", notes = "Create Pan Resource", response = PanCard.class)
+	@ApiOperation(value = "Create PanCard Resource", notes = "Create Pan Resource", response = PanCard.class)
 	public ResponseEntity<?> createPan(@Valid @RequestBody PanCard panCard, BindingResult bindingResult) {
 		ResponseEntity<?> errorMap = mapStateToError.mapStateToError(bindingResult);
 		if (errorMap != null) {
 			return errorMap;
 		}
 		PanCard panCardToDB = panCardService.savePanCardDetails(panCard);
-		return new ResponseEntity<PanCard>(panCardToDB, HttpStatus.CREATED);
+		Gson gson = new Gson();
+		String panCardResponse = gson.toJson(panCardToDB);
+		return new ResponseEntity<String>(panCardResponse, HttpStatus.CREATED);
 	}
 
 	@RequestMapping(path = "/get/{id}", method = RequestMethod.GET, produces = "application/json")
 	@ResponseBody
-	@ApiOperation(value = "/get", notes = "Retrieve Pan", produces = "applicaiton/json")
+	@ApiOperation(value = "Retrieve PanCard Resource", notes = "Retrieve Pan", response = PanCard.class)
 	public ResponseEntity<?> getPanCardByID(
 			@ApiParam(value = "id", required = true) @PathVariable(value = "id") Long id) {
 		PanCard panCardToDB = panCardService.getPanCardById(id);
-		return new ResponseEntity<PanCard>(panCardToDB, HttpStatus.OK);
+		if (panCardToDB == null) {
+			return new ResponseEntity<String>("Sorry No Data Found For This ID:", HttpStatus.BAD_REQUEST);
+		}
+		Gson gson = new Gson();
+		String stringGson = gson.toJson(panCardToDB);
+		return new ResponseEntity<String>(stringGson, HttpStatus.OK);
 	}
 
 	@RequestMapping(path = "/get", method = RequestMethod.GET, produces = "application/json")
 	@ResponseBody
-	@ApiOperation(value = "/get", notes = "Retrieve All Pan Card Details", response = PanCard.class, responseContainer = "LIST")
+	@ApiOperation(value = "Retrieve All PanCard Resource From DB", notes = "Retrieve All Pan Card Details", response = PanCard.class, responseContainer = "LIST")
 	public ResponseEntity<?> getAllPanCard() {
 		List<PanCard> panCardList = panCardService.getAllPanCardDetails();
 		if (panCardList.size() == 0 || panCardList == null) {
-			return new ResponseEntity<String>("Sorry Not Data Found In The DB", HttpStatus.CREATED);
+			return new ResponseEntity<String>("Sorry Not Data Found In The DB", HttpStatus.BAD_REQUEST);
 		}
-		return new ResponseEntity<List<PanCard>>(panCardList, HttpStatus.OK);
+		Gson gson = new Gson();
+		String response = gson.toJson(panCardList);
+
+		return new ResponseEntity<String>(response, HttpStatus.OK);
 	}
 
-	@RequestMapping(path = "/delete/{id}", method = RequestMethod.DELETE, produces = "application/json")
+	@RequestMapping(path = "/delete/{id}", method = RequestMethod.DELETE, produces = "text/plain")
 	@ResponseBody
-	@ApiOperation(value = "/delete", notes = "Remove PanCard Resource")
+	@ApiOperation(value = "Remove Resource Of PanCard From The DB", notes = "Remove PanCard Resource", response = String.class)
 	public ResponseEntity<?> deletePanById(
 			@ApiParam(value = "id", required = true) @PathVariable(value = "id") Long id) {
+
+		ResponseEntity<?> responseEntity = getPanCardByID(id);
+		String stringResponse = (String) responseEntity.getBody();
+		if (stringResponse.equals("Sorry No Data Found For This ID:")) {
+			return new ResponseEntity<String>("No Data Found For This ID:- " + id, HttpStatus.BAD_REQUEST);
+		}
 		String response = panCardService.deletePanCardByID(id);
+		if (response == null) {
+			return new ResponseEntity<String>("Sorry Could Not Found Data For This ID:", HttpStatus.BAD_REQUEST);
+		}
 		return new ResponseEntity<String>(response, HttpStatus.OK);
 	}
 }
