+++ src/main/java/com/example/demo/controller/PersonRestController.java
@@ -23,6 +23,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.demo.beans.Person;
 import com.example.demo.service.MapStateToError;
 import com.example.demo.service.PersonService;
+import com.google.gson.Gson;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -32,7 +33,7 @@ import io.swagger.annotations.ApiResponses;
 
 @RestController
 @RequestMapping(path = "/api/object")
-@Api(value = "/api/object", description = "Manage Person Resource")
+@Api(value = "Person Object Crud Operations", description = "Crud Operation Of Person Object")
 public class PersonRestController {
 
 	private static final Logger logger = LoggerFactory.getLogger(PersonRestController.class);
@@ -52,7 +53,9 @@ public class PersonRestController {
 			return errorMap;
 		}
 		Person personToDB = personService.createPerson(person);
-		return new ResponseEntity<Person>(personToDB, HttpStatus.CREATED);
+		Gson gson = new Gson();
+		String stringResponse = gson.toJson(personToDB);
+		return new ResponseEntity<String>(stringResponse, HttpStatus.CREATED);
 	}
 
 	@RequestMapping(path = "/get/{id}", method = RequestMethod.GET, produces = "application/json")
@@ -66,29 +69,43 @@ public class PersonRestController {
 		if (personFromDB == null) {
 			return new ResponseEntity<String>("Sorry No Data Exist For This ID", HttpStatus.BAD_REQUEST);
 		}
-		return new ResponseEntity<Person>(personFromDB, HttpStatus.OK);
+		Gson gson = new Gson();
+		String stringResponse = gson.toJson(personFromDB);
+		return new ResponseEntity<String>(stringResponse, HttpStatus.OK);
 	}
 
 	@RequestMapping(path = "/get", method = RequestMethod.GET, produces = "application/json")
 	@ResponseBody
 	@ResponseStatus(HttpStatus.OK)
-	@ApiOperation(value = "Retrieve Person Resource From The DB", notes = "", response = Person.class, responseContainer = "List")
+	@ApiOperation(value = "Retrieve Person Resource From The DB", notes = "", response = Person.class, responseContainer = "LIST")
 	public ResponseEntity<?> getAllPerson() {
 		List<Person> personFromDB = personService.getAllPerson();
 		if (personFromDB == null || personFromDB.size() == 0) {
 			return new ResponseEntity<String>("Sorry No Data Exist", HttpStatus.BAD_REQUEST);
 		}
-		return new ResponseEntity<List<Person>>(personFromDB, HttpStatus.OK);
+		Gson gson = new Gson();
+		String stringResponse = gson.toJson(personFromDB);
+		return new ResponseEntity<String>(stringResponse, HttpStatus.OK);
 	}
 
-	@RequestMapping(path = "/delete/{id}", method = RequestMethod.DELETE, produces = "application/json")
+	@RequestMapping(path = "/delete/{id}", method = RequestMethod.DELETE, produces = "plain/text")
 	@ResponseBody
 	@ResponseStatus(HttpStatus.OK)
-	@ApiOperation(value = "Remove Son Resource From DB Using ID", notes = "Id Is Mandatory")
+	@ApiOperation(value = "Remove Son Resource From DB Using ID", notes = "Id Is Mandatory", response = String.class)
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "SuccessFully Deleted"),
 			@ApiResponse(code = 400, message = "In Case Son Resources Could Not Be Deleted") })
 	public ResponseEntity<?> deletePersonByID(
 			@ApiParam(value = "id", required = true) @PathVariable(value = "id") @NotBlank @NotEmpty Long id) {
+
+		ResponseEntity<?> responseEntity = getPersonByID(id);
+		String res = (String) responseEntity.getBody();
+		if (res.equals("Sorry No Data Exist For This ID")) {
+			return new ResponseEntity<String>("Sorry No Data Found For This ID", HttpStatus.BAD_REQUEST);
+		}
+
+		Person person = new Gson().fromJson(res, Person.class);
+		logger.info("Person Details:-" + person);
+
 		String response = personService.deletePersonByID(id);
 		if (response == null) {
 			return new ResponseEntity<String>("Sorry No Data Exist For This Id", HttpStatus.BAD_REQUEST);
